Implement script to wait for server to become online
When installing keycloak using helm, the post-install hooks will be executed
as soon as helm is finished. However, at that moment, the keycloak webserver
is still starting and probably waiting for postgres to start.

As the scripts are dependent on keycloak being started, it will wait for max
5 minutes until the server is up.

See also https://github.com/kubernetes/helm/issues/3282